ChairManager: ConfigManager: Move to Common
