+++ CommonChairManager/Manager/ConfigManager.h
