+++ CommonChairManager/Manager/ModConfig.h
